# perf: [NPM] Use Equal function for label comparison instead of using reflect.DeepEqual in namespaceController