<!-- Thank you for helping Azure Container Networking with a pull request!
Use conventional commit messages, such as
  feat: add a knob to the frobnitz
or
  fix: repair hole in wumpus
And read this for faster PR reviews: https://github.com/kubernetes/community/blob/master/contributors/guide/pull-requests.md#best-practices-for-faster-reviews -->

**Reason for Change**:
<!-- What does this PR improve or fix in Azure Container Networking? -->
To avoid using heavy deepEqual to compare label and be consistent with other controller codes.

**Issue Fixed**:
<!-- If this PR fixes GitHub issue 1234, add "Fixes #1234" to the next line. -->


**Requirements**:
<!-- Put an "X" character inside the brackets of each completed task. Some may be optional depending on the PR. -->


- [x] uses [conventional commit messages](https://www.conventionalcommits.org/)
  <!-- Common commit types:
        build: Build 🏭
        chore: Maintenance 🔧
        ci: Continuous Integration 💜
        docs: Documentation 📘
        feat: Features 🌈
        fix: Bug Fixes 🐞
        perf: Performance Improvements 🚀
        refactor: Code Refactoring 💎
        revert: Revert Change ◀️
        style: Code Style 🎶
        security: Security Fix 🛡️
        test: Testing 💚 -->
- [ ] includes documentation
- [ ] adds unit tests


**Notes**:
Test two functions and their results.
```go
package testing

import (
	"reflect"
	"testing"

	k8slabels "k8s.io/apimachinery/pkg/labels"
)

func BenchmarkLabelEqual(b *testing.B) {
	labelA := map[string]string{
		"project": "myproject",
		"role":    "frontend",
	}
	labelB := map[string]string{
		"project": "myproject",
		"role":    "frontend",
	}
	for i := 0; i < b.N; i++ {
		k8slabels.Equals(labelA, labelB)
	}
}

func BenchmarkLabelDeepEqual(b *testing.B) {
	labelA := map[string]string{
		"project": "myproject",
		"role":    "frontend",
	}
	labelB := map[string]string{
		"project": "myproject",
		"role":    "frontend",
	}
	for i := 0; i < b.N; i++ {
		reflect.DeepEqual(labelA, labelB)
	}
}

```
```shell
Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -bench ^BenchmarkLabelEqual$ benchmarking.com

goos: linux
goarch: amd64
pkg: benchmarking.com
cpu: Intel(R) Core(TM) i7-1065G7 CPU @ 1.30GHz
BenchmarkLabelEqual-8   	18292081	        70.08 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	benchmarking.com	1.360s


Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -bench ^BenchmarkLabelDeepEqual$ benchmarking.com
goos: linux
goarch: amd64
pkg: benchmarking.com
cpu: Intel(R) Core(TM) i7-1065G7 CPU @ 1.30GHz
BenchmarkLabelDeepEqual-8   	 1522663	      1082 ns/op	     240 B/op	       8 allocs/op
PASS
ok  	benchmarking.com	3.529s
```