@JungukCho 2021-11-16
> 😍

Thank you for cool reference for `equal` function